pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "content-engine"
version = "0.1.0"
description = "Short-form content engine for social media monetization"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.0",
    "openai-whisper>=20231117",
    "ffmpeg-python>=0.2.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "anthropic>=0.40.0",
    "opencv-python-headless>=4.8",
    "mediapipe>=0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
content-engine = "content_engine.cli:cli"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
